requirements.txt
Flask
Flask-Moment
openai
idna
python-dotenv
Werkzeug
requests
numpy
pydub
pyannote.audio
torch
torchaudio
PyMuPDF
python-docx
python-pptx
PyJWT